fixed cc_reshare from uchar to int
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@2682 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ globals.h
@@ -634,7 +634,7 @@ struct s_reader  //contains device info, reader info and card info
 	BIGNUM ucpk;
 	////variables from reader-viaccess.c 
 	struct geo_cache last_geo;
-	uchar cc_reshare;
+	int cc_reshare;
 	int lb_weight; //loadbalance weight factor, if unset, weight=100. The higher the value, the higher the usage-possibility
 };
